+++ HISTORY.rst
@@ -10,7 +10,7 @@ History
 0.2.3 (2018-05-14)
 ------------------
 
-- Fix lookup_index functionality for composite fields. 
+- Fix lookup_index functionality for composite fields.
 
 - Make model name property optional. That is, 'unknown' will be returned if the
   name cannot be derived.